simplifying stream creation with << and >> operators, working on graph optimization
