+++ tributary/streaming/calculations/rolling.py
@@ -1,5 +1,5 @@
 from .utils import _CALCULATIONS_GRAPHVIZSHAPE
-from ..base import Node
+from ..node import Node
 
 
 def Count(node):
@@ -14,8 +14,7 @@ def Count(node):
 
     ret = Node(foo=foo, foo_kwargs=None, name='Count', inputs=1, graphvizshape=_CALCULATIONS_GRAPHVIZSHAPE)
     ret._count = 0
-    node._downstream.append((ret, 0))
-    ret._upstream.append(node)
+    node >> ret
     return ret
 
 
@@ -31,9 +30,7 @@ def Max(node):
 
     ret = Node(foo=foo, foo_kwargs=None, name='Max', inputs=1, graphvizshape=_CALCULATIONS_GRAPHVIZSHAPE)
     ret._max = None
-
-    node._downstream.append((ret, 0))
-    ret._upstream.append(node)
+    node >> ret
     return ret
 
 
@@ -49,8 +46,7 @@ def Min(node):
 
     ret = Node(foo=foo, foo_kwargs=None, name='Min', inputs=1, graphvizshape=_CALCULATIONS_GRAPHVIZSHAPE)
     ret._min = None
-    node._downstream.append((ret, 0))
-    ret._upstream.append(node)
+    node >> ret
     return ret
 
 
@@ -75,9 +71,7 @@ def Sum(node):
 
     ret = Node(foo=foo, foo_kwargs=None, name='Sum', inputs=1, graphvizshape=_CALCULATIONS_GRAPHVIZSHAPE)
     ret._sum = 0
-
-    node._downstream.append((ret, 0))
-    ret._upstream.append(node)
+    node >> ret
     return ret
 
 
@@ -103,8 +97,7 @@ def Average(node):
     ret._sum = 0
     ret._count = 0
 
-    node._downstream.append((ret, 0))
-    ret._upstream.append(node)
+    node >> ret
     return ret
