+++ tributary/streaming/output/http.py
@@ -1,7 +1,7 @@
 import requests
 import json as JSON
 from .output import _OUTPUT_GRAPHVIZSHAPE
-from ..base import Node
+from ..node import Node
 from ...base import StreamEnd
 
 
@@ -45,6 +45,5 @@ def HTTP(node, url='', json=False, wrap=False, field=None, proxies=None, cookies
         return msg
 
     ret = Node(foo=_send, name='Http', inputs=1, graphvizshape=_OUTPUT_GRAPHVIZSHAPE)
-    node._downstream.append((ret, 0))
-    ret._upstream.append(node)
+    node >> ret
     return ret
