+++ tributary/streaming/output/socketio.py
@@ -2,7 +2,7 @@ import json as JSON
 from socketIO_client_nexus import SocketIO as SIO
 from urllib.parse import urlparse
 from .output import _OUTPUT_GRAPHVIZSHAPE
-from ..base import Node
+from ..node import Node
 
 
 def SocketIO(node, url, channel='', field='', sendinit=None, json=False, wrap=False, interval=1):
@@ -39,6 +39,5 @@ def SocketIO(node, url, channel='', field='', sendinit=None, json=False, wrap=Fa
         return data
 
     ret = Node(foo=_sio, name='SocketIO', inputs=1, graphvizshape=_OUTPUT_GRAPHVIZSHAPE)
-    node._downstream.append((ret, 0))
-    ret._upstream.append(node)
+    node >> ret
     return ret
